requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2
pyyaml==6.0.1
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
redis==5.0.1
celery==5.3.4
structlog==23.2.0
prometheus-client==0.19.0